Cleanup, delete and organise files into folders
