+++ LBHTenancyAPI/Controllers/BaseController.cs
@@ -1,4 +1,5 @@
 using LBHTenancyAPI.Extensions.Controller;
+using LBHTenancyAPI.Infrastructure.UseCase.Execution;
 using LBHTenancyAPI.UseCases.ArrearsAgreements;
 using Microsoft.AspNetCore.Mvc;
