+++ LBHTenancyAPI/Extensions/Validation/APIErrorExtensions.cs
@@ -1,6 +1,7 @@
 using System.Collections.Generic;
 using FluentValidation.Results;
 using LBHTenancyAPI.Infrastructure.API;
+using LBHTenancyAPI.Infrastructure.UseCase.Execution;
 
 namespace LBHTenancyAPI.Extensions.Validation
 {
