--- LBHTenancyAPI/Gateways/IPaymentTransactionsGateway.cs
@@ -1,11 +1,0 @@
-using System.Collections.Generic;
-using System.Threading.Tasks;
-
-namespace LBHTenancyAPI.Gateways
-{
-    public interface IPaymentTransactionsGateway
-    {
-        string GetTransactionDescription(string transactionType);
-
-    }
-}
