+++ LBHTenancyAPI/Infrastructure/API/APIError.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using AgreementService;
 using LBHTenancyAPI.Infrastructure.API;
+using LBHTenancyAPI.Infrastructure.UseCase.Execution;
 
 namespace LBHTenancyAPI.UseCases.ArrearsAgreements
 {
