+++ LBHTenancyAPI/Infrastructure/API/IRequest.cs
@@ -1,0 +1,16 @@
+namespace LBHTenancyAPI.UseCases.ArrearsAgreements
+{
+    /// <summary>
+    /// Request Interface
+    /// </summary>
+    public interface IRequest
+    {
+        /// <summary>
+        /// Encapsulates the validation by making the request responsible for validating itself
+        /// </summary>
+        /// <typeparam name="T"></typeparam>
+        /// <param name="request"></param>
+        /// <returns></returns>
+        RequestValidationResponse Validate<T>(T request);
+    }
+}
