+++ LBHTenancyAPI/Infrastructure/Logging/SentryLogger.cs
