+++ LBHTenancyAPI/LBHTenancyAPI.csproj
@@ -8,6 +8,7 @@
       <NoWarn>$(NoWarn);1591</NoWarn>
   </PropertyGroup>
   <ItemGroup>
+    <Folder Include="Interfaces\" />
     <Folder Include="wwwroot\" />
   </ItemGroup>
   <ItemGroup>
