+++ LBHTenancyAPI/UseCases/ArrearsAgreements/CreateArrearsAgreementUseCase.cs
@@ -3,6 +3,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using AgreementService;
 using LBHTenancyAPI.Gateways.Arrears;
+using LBHTenancyAPI.Infrastructure.UseCase.Execution;
 
 namespace LBHTenancyAPI.UseCases.ArrearsAgreements
 {
