--- LBHTenancyAPI/UseCases/ArrearsAgreements/IRequest.cs
@@ -1,7 +1,0 @@
-namespace LBHTenancyAPI.UseCases.ArrearsAgreements
-{
-    public interface IRequest
-    {
-        RequestValidationResponse Validate<T>(T request);
-    }
-}
